+++ ulmo/webpage_dynamic/os_portal.py
@@ -28,27 +28,11 @@ from bokeh.events import DoubleTap, PinchEnd, PanEnd, Reset
 from bokeh.tile_providers import get_provider, Vendors
 from bokeh.transform import linear_cmap
 
+from ulmo.webpage_dynamic import utils as portal_utils
 
 from IPython import embed
 
-def mercator_coord(lat, lon):
-    """Function to switch from lat/long to mercator coordinates
-
-    Args:
-        lat (float or np.ndarray): [description]
-        lon (float or np.ndarray): [description]
-
-    Returns:
-        tuple: x, y values in mercator
-    """
-    r_major = 6378137.000
-    x = r_major * np.radians(lon)
-    scale = x/lon
-    y = 180.0/np.pi * np.log(np.tan(np.pi/4.0 + 
-        lat * (np.pi/180.0)/2.0)) * scale
-    return x, y
-
-class os_web(object):
+class OSPortal(object):
     """ Primary class for the web site """
     def __init__(self, data_dict, config=None, verbose=False): #images, obj_ids, metric_dict, umap_data):
 
@@ -366,7 +350,7 @@ class os_web(object):
             self.ylim_all[umap] = temp_ylim
         '''
 
-        points = get_decimated_region_points(self.xlim[0], self.xlim[1], self.ylim[0], self.ylim[1],
+        points = portal_utils.get_decimated_region_points(self.xlim[0], self.xlim[1], self.ylim[0], self.ylim[1],
                                              self.umap_source.data, self.DECIMATE_NUMBER)
 
         self.umap_source_view = ColumnDataSource(
@@ -812,7 +796,7 @@ class os_web(object):
 
         # Geo map
         if self.geo:
-            mercator_x, mercator_y =  mercator_coord(
+            mercator_x, mercator_y =  portal_utils.mercator_coord(
                 np.array(tdict['lat']), np.array(tdict['lon']))
             #print(mercator_y)
             geo_dict = dict(mercator_x=mercator_x.tolist(), 
@@ -976,7 +960,7 @@ class os_web(object):
                  (py_start > np.min(uy) ) or
                  (py_end   < np.max(uy) )   ):
 
-                background_objects = get_decimated_region_points(
+                background_objects = portal_utils.get_decimated_region_points(
                     self.umap_figure.x_range.start,
                     self.umap_figure.x_range.end,
                     self.umap_figure.y_range.start,
@@ -1006,7 +990,7 @@ class os_web(object):
 
         #embedding = self.dropdown_dict['embedding']
         print('get_new_view_keep_selected')
-        _, _, is_relevant = get_relevant_objects_coords(self.umap_source.data)
+        _, _, is_relevant = portal_utils.get_relevant_objects_coords(self.umap_source.data)
         selected_objects = [s for s in selected_objects_ if is_relevant[int(s)]]
         selected_objects = np.array(selected_objects)
         background_objects = np.array(background_objects)
@@ -1101,7 +1085,7 @@ class os_web(object):
             if self.verbose:
                 print('reset double tap')
 
-            background_objects = get_decimated_region_points(
+            background_objects = portal_utils.get_decimated_region_points(
                 self.xlim[0],
                 self.xlim[1],
                 self.ylim[0],
@@ -1178,184 +1162,3 @@ class os_web(object):
 
         return
 
-
-def get_region_points(x_min, x_max, y_min, y_max, datasource):
-    """ Get the points within the box
-
-    Args:
-        x_min ([type]): [description]
-        x_max ([type]): [description]
-        y_min ([type]): [description]
-        y_max ([type]): [description]
-        datasource ([type]): [description]
-
-    Returns:
-        [type]: [description]
-    """
-    IGNORE_TH = -9999
-    xs = np.array(datasource['xs'])
-    ys = np.array(datasource['ys'])
-    cd = datasource['color_data']
-    nof_objects = len(cd)
-    if True:
-        is_in_box = np.logical_and.reduce([xs >= x_min, xs <= x_max, ys >= y_min, ys <= y_max, ys > IGNORE_TH, xs > IGNORE_TH])
-    else:
-        is_in_box = np.logical_and.reduce([xs >= x_min, xs <= x_max, ys >= y_min, ys <= y_max])
-    return np.where(is_in_box)[0]
-
-
-def get_relevant_objects_coords(datasource):
-    """ ??
-
-    Args:
-        datasource ([type]): [description]
-
-    Returns:
-        [type]: [description]
-    """
-    IGNORE_TH = -999
-    xs = np.array(datasource['xs'])
-    ys = np.array(datasource['ys'])
-    relevant_objects = np.logical_and.reduce([ys > IGNORE_TH, xs > IGNORE_TH])
-    return xs[relevant_objects], ys[relevant_objects], relevant_objects
-
-
-def get_decimated_region_points(x_min, x_max, y_min, y_max, datasource, DECIMATE_NUMBER):
-    """ ??
-
-    Args:
-        x_min ([type]): [description]
-        x_max ([type]): [description]
-        y_min ([type]): [description]
-        y_max ([type]): [description]
-        datasource ([type]): [description]
-        DECIMATE_NUMBER ([type]): [description]
-
-    Returns:
-        [type]: [description]
-    """
-    is_in_box_inds = get_region_points(x_min, x_max, y_min, y_max, datasource)
-    print('total points before decimation', len(is_in_box_inds))
-    if len(is_in_box_inds) < DECIMATE_NUMBER:
-        return is_in_box_inds
-    random_objects_ = np.random.choice(is_in_box_inds, DECIMATE_NUMBER, replace=False)
-    random_objects = [datasource['names'][r] for r in random_objects_]
-    return random_objects
-
-
-
-# CODE HERE AND DOWN IS FOR TESTING
-def get_test_os_session(doc):
-    images, objids, metric, umapd = grab_dum_data()
-    # Instantiate
-    sess = os_web(images, objids, metric, umapd)
-    return sess(doc)
-
-def get_modis_subset_os_session(doc):
-    data_dict = grab_modis_subset()
-    # Instantiate
-    sess = os_web(data_dict)
-    return sess(doc)
-
-def grab_dum_data():
-    nobj = 100
-    dum_images = np.random.uniform(size=(nobj, 64, 64))
-    dum_objids = np.arange(nobj)
-    dum_umap = np.random.uniform(size=(nobj,2)) 
-    dum_LL = np.random.uniform(low=0., high=100., size=nobj)
-    dum_metric = dict(LL=dum_LL)
-    dum_umapd = dict(UMAP=dum_umap)
-    embed(header='NEED TO REFACTOR')
-    #
-    return dum_images, dum_objids, dum_metric, dum_umapd
-
-def grab_modis_subset():
-    # Load up 
-    sst_dir='/data/Projects/Oceanography/AI/OOD/SST/MODIS_L2/PreProc/'
-    data_file = os.path.join(sst_dir, 
-                             'MODIS_R2019_2010_95clear_128x128_preproc_std.h5')
-    results_path = '/data/Projects/Oceanography/AI/SSL/portal/'
-    umaps_path = os.path.join(results_path, 'embeddings')
-
-    # Images
-    nimgs = 100000
-    sub_idx = np.arange(nimgs)
-    f = h5py.File(data_file, 'r') 
-    images = f["valid"][sub_idx,0,:,:]
-    f.close()
-
-    # UMAP
-    umap_file = os.path.join(umaps_path, 'UMAP_2010_valid_v1.npz')
-    f = np.load(umap_file, allow_pickle=False)
-    e1, e2 = f['e1'], f['e2']
-
-    # Metrics
-    results_file = os.path.join(results_path, 'ulmo_2010_valid_v1.parquet')
-    res = pandas.read_parquet(results_file)
-    metric_dict = {'LL': res.LL.values[sub_idx], 
-                   'lat': res.lat.values[sub_idx],
-                   'lon': res.lon.values[sub_idx],
-                   'avgT': res.mean_temperature.values[sub_idx],
-                   'DT': (res.T90-res.T10).values[sub_idx],
-                   'obj_ID': sub_idx,
-    }
-
-    # Repack
-    data_dict = {
-        'images': images,
-        'xy_scatter': dict(UMAP=(np.array([e1, e2]).T)[sub_idx]),
-        'metrics': metric_dict,
-    }
-
-    return data_dict
-
-
-def main(flg):
-    flg = int(flg)
-
-    # Deprecated test
-    if flg & (2 ** 0):
-        pass
-    
-    # Test main class
-    if flg & (2 ** 1):
-        dum_images, dum_objids, dum_metric, dum_umapd = grab_dum_data()
-        sess = os_web(dum_images, dum_objids, dum_metric, dum_umapd)
-        print("Success!")
-
-    # Real deal
-    if flg & (2 ** 2):
-        server = Server({'/': get_test_os_session}, num_procs=1)
-        server.start()
-        print('Opening Bokeh application for test data on http://localhost:5006/')
-
-        server.io_loop.add_callback(server.show, "/")
-        server.io_loop.start()
-
-    # Test modis subset
-    if flg & (2 ** 3):
-        data_dict = grab_modis_subset()
-        sess = os_web(data_dict)
-
-    if flg & (2 ** 4):
-        server = Server({'/': get_modis_subset_os_session}, num_procs=1)
-        server.start()
-        print('Opening Bokeh application for MODIS subset on http://localhost:5006/')
-
-        server.io_loop.add_callback(server.show, "/")
-        server.io_loop.start()
-
-
-if __name__ == '__main__':
-    import sys
-    if len(sys.argv) == 1:
-        flg = 0
-        #flg += 2 ** 0  # Test bokeh
-        #flg += 2 ** 1  # Test object
-        #flg += 2 ** 2  # Full Test 
-        #flg += 2 ** 3  # Test load MODIS subset
-        flg += 2 ** 4  # Full MODIS subset
-    else:
-        flg = sys.argv[1]
-
-    main(flg)
